[feat]멀티키워드 생성 req,res dto 변경

+++ src/main/java/com/hana4/keywordhanaro/model/entity/keyword/MultiKeyword.java
@@ -47,5 +47,5 @@ public class MultiKeyword {
 	private Keyword keyword;
 
 	@Column(name = "seqOrder", nullable = false)
-	private byte seqOrder;
+	private Long seqOrder;
 }
